feat: enhance interface detail
At a glance the type structure of `is` together with `isOptions` wasn’t very explicit. Interfaces have now been broken out into their own file to show a little bit more how the model of `isOptions` is formed with the added benefit of making the `is` type more clear
